+++ src/iarray_operator.c
@@ -420,10 +420,12 @@ static ina_rc_t _iarray_gemv(iarray_context_t *ctx, iarray_container_t *a, iarra
         // Make blocks multiplication
         switch (dtype) {
             case IARRAY_DATA_TYPE_DOUBLE:
-                cblas_dgemv(CblasRowMajor, flag_a, cB0, cB1, 1.0, (double *) a_block, ld_a, (double *) b_block, 1, 1.0, (double *) c_block, 1);
+                cblas_dgemv(CblasRowMajor, flag_a, (int) cB0, (int) cB1, 1.0, (double *) a_block,
+                            ld_a, (double *) b_block, 1, 1.0, (double *) c_block, 1);
                 break;
             case IARRAY_DATA_TYPE_FLOAT:
-                cblas_sgemv(CblasRowMajor, flag_a, cB0, cB1, 1.0f, (float *) a_block, ld_a, (float *) b_block, 1, 1.0f, (float *) c_block, 1);
+                cblas_sgemv(CblasRowMajor, flag_a, (int) cB0, (int) cB1, 1.0f, (float *) a_block,
+                            ld_a, (float *) b_block, 1, 1.0f, (float *) c_block, 1);
                 break;
             default:
                 IARRAY_TRACE1(iarray.error, "The data type is invalid");
